packages/agent-sdk/pyproject.toml
[project]
name = "abenix-sdk"
version = "0.1.0"
description = "Python SDK for building Abenix agents"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
